requirements.txt
fastapi[all]
python-dotenv
pydantic-settings
openai
google-generativeai
anthropic
